Cambios en el modelo de producto y en la vista y controlador de detalles del producto...correcion error que no dejaba visualizar correctamente

+++ lib/app/modules/product_details/controllers/product_details_controller.dart
@@ -128,7 +128,7 @@ class ProductDetailsController extends GetxController {
 }
 
 //getratingString
-String getRatingString(int rating) {
+String getRatingString(double rating) {
     if (rating == 0.0) {
       return 'Sin calificación';
     } else if (rating == 1.0) {
